refactored App to have useEffect and axios, setting state and passing it to MovieList, have to add ...props
