+++ client/src/App.js
@@ -1,5 +1,6 @@
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import {Route} from 'react-router-dom'
+import axios from 'axios';
 
 import SavedList from './Movies/SavedList';
 import MovieList from "./Movies/MovieList";
@@ -7,6 +8,23 @@ import Movie from "./Movies/Movie"
 
 const App = () => {
   const [savedList, setSavedList] = useState( [] );
+
+  const [movies, setMovies] = useState([])
+  
+  useEffect(() => {
+    const getMovies = () => {
+      axios
+        .get('http://localhost:5000/api/movies')
+        .then(response => {
+          setMovies(response.data);
+        })
+        .catch(error => {
+          console.error('Server Error', error);
+        });
+    }
+    
+    getMovies();
+  }, []);
 
   const addToSavedList = movie => {
     setSavedList( [...savedList, movie] );
@@ -18,7 +36,7 @@ const App = () => {
       
      
 
-      <Route exact path="/" render={() => <MovieList/>}/>
+      <Route exact path="/" render={() => <MovieList movies={movies}/>}/>
       <Route path="/movies/:id" render={()=> <Movie/>}/>
     </div>
   );
