+++ client/src/Movies/MovieList.js
@@ -4,25 +4,25 @@ import axios from 'axios';
 import {Link} from "react-router-dom"
 
 const MovieList = props => {
-  const [movies, setMovies] = useState([])
-  useEffect(() => {
-    const getMovies = () => {
-      axios
-        .get('http://localhost:5000/api/movies')
-        .then(response => {
-          setMovies(response.data);
-        })
-        .catch(error => {
-          console.error('Server Error', error);
-        });
-    }
+  // const [movies, setMovies] = useState([])
+  // useEffect(() => {
+  //   const getMovies = () => {
+  //     axios
+  //       .get('http://localhost:5000/api/movies')
+  //       .then(response => {
+  //         setMovies(response.data);
+  //       })
+  //       .catch(error => {
+  //         console.error('Server Error', error);
+  //       });
+  //   }
     
-    getMovies();
-  }, []);
+  //   getMovies();
+  // }, []);
   
   return (
     <div className="movie-list">
-      {movies.map(movie => (
+      {props.movies.map(movie => (
         <MovieDetails key={movie.id} movie={movie} />
       ))}
